+++ src/modules/projects/projects.controller.ts
@@ -20,11 +20,17 @@ import {
 import { FileInterceptor } from '@nestjs/platform-express';
 import { imageFileAvatarFilter } from '@/cloudinary/filter/filter.user.avatar';
 import { MulterFile } from '@/types/multer-file';
+import { Roles } from '@/common/decorators/roles.decorator';
+import { Role } from '@/enums/role.enum';
+import { CurrentUser } from '@/common/decorators/current-user.decorator';
+import { JwtPayload } from '@/types/jwt-payload.interface';
+import { checkPermission } from '@/common/utils/role/auth-utils';
 
 @Controller('projects')
 export class ProjectsController {
   constructor(private readonly projectsService: ProjectsService) {}
 
+  @Roles(Role.ADMIN, Role.LEADER)
   @Post('create')
   @UseInterceptors(
     FileInterceptor('file', {
@@ -39,13 +45,26 @@ export class ProjectsController {
     return this.projectsService.createProjet(createProjectDto, file);
   }
 
+  @Roles(Role.ADMIN, Role.LEADER)
   @Get('pagination')
   Pagination(@Query() paginationDto: PaginationProjectDto) {
     return this.projectsService.Pagination(paginationDto);
   }
 
   @Get()
-  findOne(@Query('id') id: string) {
+  async findOne(@Query('id') id: string, @CurrentUser() user: JwtPayload) {
+    /**
+     * Kiểm tra quyền truy cập với role và id
+     * @param user Thông tin user hiện tại (jwt payload)
+     * @param allowedRoles Các role được phép thao tác không cần so sánh id
+     * @param checkOwnUser Nếu true, user chỉ được thao tác với chính mình nếu không thuộc allowedRoles
+     */
+    const isProject = await this.projectsService.verifyUserAccessToProjectId(
+      id,
+      user.id,
+    );
+    checkPermission(user, isProject, [Role.ADMIN]);
+
     return this.projectsService.findProjectById(id);
   }
 
@@ -56,16 +75,30 @@ export class ProjectsController {
       fileFilter: imageFileAvatarFilter,
     }),
   )
-  update(
+  async update(
     @Query('id') id: string,
     @Body() updateProjectDto: UpdateProjectDto,
+    @CurrentUser() user: JwtPayload,
     @UploadedFile() file?: MulterFile,
   ) {
+    // Kiểm tra quyền truy cập với role và id
+    const hasAccess = await this.projectsService.verifyUserAccessToProjectId(
+      id,
+      user.id,
+    );
+    checkPermission(user, hasAccess, [Role.ADMIN]);
+
     return this.projectsService.updateProject(id, updateProjectDto, file);
   }
 
   @Delete('delete')
-  remove(@Query('id') id: string) {
+  async remove(@Query('id') id: string, @CurrentUser() user: JwtPayload) {
+    // Kiểm tra quyền truy cập với role và id
+    const hasAccess = await this.projectsService.verifyUserAccessToProjectId(
+      id,
+      user.id,
+    );
+    checkPermission(user, hasAccess, [Role.ADMIN]);
     return this.projectsService.removeProjectDev(id);
   }
 }
